fetch url for downloading in background process
Moves code that fetches the url into the background process
that downloads the chunk. That way we use it immediately
after we retrieve it. Otherwise there is a possiblity
that we end up trying to download from an expired url.
This scenario was likely when retrying downloads due
to connection problems.