Add leading slash to remote paths
In support of #243 adds leading slash to remote paths.
This allows users to naturally specify the top level 'directory'
of a project with '/'. Previously this was omitted to match
the DukeDS web portal.

Changes the download command arguments include/exclude to allow
leading slash or the current functionality of omitting the leading
slash.